mvoe utxo input selector to separate module
