+++ rust/frameworks/tw_utxo/src/sighash_computer/mod.rs
@@ -31,6 +31,7 @@ pub struct UtxoToSign {
 
 /// Transaction preimage arguments.
 /// It provides required options to sign each UTXO of the transaction.
+// TODO: Move this to another module.
 pub struct TxSigningArgs {
     pub utxos_to_sign: Vec<UtxoToSign>,
     pub sighash_ty: SighashType,
@@ -74,6 +75,7 @@ pub struct UtxoSighash {
     pub sighash: Data,
 }
 
+// TODO: Move this to another module.
 pub struct SpendingData {
     pub script_sig: Script,
     pub witness: Witness,
@@ -180,25 +182,6 @@ where
     /// Consider using [`SighashComputer::verify_signatures`] before calling [`SighashComputer::compile`]
     /// if the signatures were computed externally.
     pub fn compile(mut self, claims: Vec<SpendingData>) -> UtxoResult<Transaction> {
-        self.do_compile(claims)?;
-
-        Ok(self.transaction_to_sign)
-    }
-
-    pub fn input_selector(
-        mut self,
-        claims: Vec<SpendingData>,
-    ) -> UtxoResult<SelectionBuilder<Transaction>> {
-        self.do_compile(claims)?;
-
-        Ok(SelectionBuilder {
-            transaction_to_sign: self.transaction_to_sign,
-            args: self.args,
-            _phantom: PhantomData,
-        })
-    }
-
-    fn do_compile(&mut self, claims: Vec<SpendingData>) -> UtxoResult<()> {
         // There should be the same number of UTXOs and their meta data.
         if self.args.utxos_to_sign.len() != self.transaction_to_sign.inputs().len() {
             return Err(UtxoError(UtxoErrorKind::Error_internal));
@@ -220,155 +203,6 @@ where
             utxo.set_witness(claim.witness);
         }
 
-        Ok(())
-    }
-}
-
-#[non_exhaustive]
-pub enum InputSelector {
-    Ascending,
-    Descending,
-    InOrder,
-}
-
-pub struct SelectionBuilder<Transaction> {
-    transaction_to_sign: Transaction,
-    args: TxSigningArgs,
-    _phantom: PhantomData<Transaction>,
-}
-
-impl<Transaction> SelectionBuilder<Transaction>
-where
-    Transaction: TransactionPreimage + TransactionInterface + TransactionFee,
-    Transaction::Output: TxOutputInterface,
-{
-    pub fn select_inputs(mut self, selector: InputSelector, fee_rate: Amount) -> UtxoResult<Self> {
-        // Calculate the total output amount.
-        let total_out = self
-            .transaction_to_sign
-            .outputs()
-            .iter()
-            .map(|output| output.value())
-            .sum::<Amount>();
-
-        debug_assert_eq!(
-            self.args.utxos_to_sign.len(),
-            self.transaction_to_sign.inputs().len()
-        );
-
-        // We clone transaction so that the orginal remains unchanged in case we
-        // return an error early.
-        let mut tx = self.transaction_to_sign.clone();
-
-        // Prepare the available UTXOs.
-        let mut utxos: Vec<(Transaction::Input, &UtxoToSign)> = tx
-            .inputs()
-            .into_iter()
-            .cloned()
-            .zip(self.args.utxos_to_sign.iter())
-            .collect::<Vec<_>>();
-
-        // Sort the UTXOs.
-        match selector {
-            InputSelector::InOrder => {
-                // Nothing to do.
-            },
-            InputSelector::Ascending => {
-                utxos.sort_by(|(_, a), (_, b)| a.amount.partial_cmp(&b.amount).unwrap());
-            },
-            InputSelector::Descending => {
-                utxos.sort_by(|(_, a), (_, b)| b.amount.partial_cmp(&a.amount).unwrap());
-            },
-        }
-
-        // Select the UTXOs to cover all the outputs and the fee.
-        let mut total_in = Amount::from(0);
-        let mut selected_utxo = Vec::with_capacity(utxos.len());
-        let mut selected_args = Vec::with_capacity(utxos.len());
-
-        let mut total_covered = false;
-        for (input, arg) in utxos {
-            // Spending script MUST be set.
-            debug_assert!(input.has_witness() || input.has_script_sig());
-
-            // Update total input amount.
-            total_in += arg.amount;
-
-            // Track the selected UTXOs.
-            selected_utxo.push(input);
-            selected_args.push(arg.clone());
-
-            // Update the transaction with (all) the selected in UTXOs.
-            tx.replace_inputs(selected_utxo.clone());
-
-            // Check if the total input amount covers the total output amount
-            // and the fee.
-            if total_in >= total_out + tx.fee(fee_rate) {
-                // Update self with selected in UTXOs.
-                tx.replace_inputs(selected_utxo);
-
-                total_covered = true;
-                break;
-            }
-        }
-
-        if !total_covered {
-            // Insufficient funds.
-            todo!()
-        }
-
-        // Update the transaction with the selected inputs.
-        self.transaction_to_sign = tx;
-        self.args.utxos_to_sign = selected_args;
-
-        Ok(self)
-    }
-
-    pub fn change(&self, fee_rate: Amount) -> UtxoResult<Amount> {
-        let change_wo_fee = self.change_without_fee()?;
-        let fee = self.transaction_to_sign.fee(fee_rate);
-
-        if fee > change_wo_fee {
-            return Err(UtxoError(UtxoErrorKind::Error_internal));
-        }
-
-        Ok(change_wo_fee - fee)
-    }
-
-    pub fn change_without_fee(&self) -> UtxoResult<Amount> {
-        let total_input = self
-            .args
-            .utxos_to_sign
-            .iter()
-            .map(|utxo| utxo.amount)
-            .sum::<Amount>();
-
-        let total_output = self
-            .transaction_to_sign
-            .outputs()
-            .iter()
-            .map(|out| out.value())
-            .sum::<Amount>();
-
-        if total_output > total_input {
-            return Err(UtxoError(UtxoErrorKind::Error_internal));
-        }
-
-        let change = total_input - total_output;
-        Ok(change)
-    }
-
-    pub fn set_change_output(mut self, change: Amount) -> UtxoResult<SighashComputer<Transaction>> {
-        let did_change = self.transaction_to_sign.set_change_amount(change);
-        if !did_change {
-            // TODO: Should this just panic?
-            return Err(UtxoError(UtxoErrorKind::Error_internal));
-        }
-
-        Ok(SighashComputer {
-            transaction_to_sign: self.transaction_to_sign,
-            args: self.args,
-            _phantom: PhantomData,
-        })
+        Ok(self.transaction_to_sign)
     }
 }
